+++ magda-web-client/src/Components/Header/Header.js
@@ -30,6 +30,7 @@ class Header extends Component {
     }
 
     render() {
+        const applicationName = this.props.strings.applicationName;
         return (
             <div className="header">
                 <div className="au-header">
@@ -40,14 +41,14 @@ class Header extends Component {
                                     <Small>
                                         <img
                                             src={config.headerMobileLogoUrl}
-                                            alt={config.appName}
+                                            alt={applicationName}
                                             className="au-header__logo"
                                         />
                                     </Small>
                                     <Medium>
                                         <img
                                             src={config.headerLogoUrl}
-                                            alt={config.appName}
+                                            alt={applicationName}
                                             className="au-header__logo"
                                         />
                                     </Medium>
@@ -121,7 +122,7 @@ const mapStateToProps = state => {
             auth: {}
         });
     }
-    return { headerNavigation };
+    return { headerNavigation, strings: state.content.strings };
 };
 
 export default connect(mapStateToProps)(Header);