+++ magda-web-client/src/Components/RecordHandler.js
@@ -10,7 +10,6 @@ import {
     fetchDistributionFromRegistry,
     resetFetchRecord
 } from "../actions/recordActions";
-import { config } from "../config";
 import defined from "../helpers/defined";
 import { gapi } from "../analytics/ga";
 import ErrorHandler from "./ErrorHandler";
@@ -364,7 +363,8 @@ class RecordHandler extends React.Component {
                                             Created{" "}
                                             <span itemProp="dateCreated">
                                                 {this.props.dataset.issuedDate}
-                                            </span>&nbsp;
+                                            </span>
+                                            &nbsp;
                                         </span>
                                     )}
 
@@ -401,14 +401,16 @@ class RecordHandler extends React.Component {
                                             />
                                         </Medium>
                                     </div>
-                                    <div className="quality-rating-box">
-                                        <QualityIndicator
-                                            quality={
-                                                this.props.dataset
-                                                    .linkedDataRating
-                                            }
-                                        />
-                                    </div>
+                                    {this.props.dataset.linkedDataRating && (
+                                        <div className="quality-rating-box">
+                                            <QualityIndicator
+                                                quality={
+                                                    this.props.dataset
+                                                        .linkedDataRating
+                                                }
+                                            />
+                                        </div>
+                                    )}
                                     {this.props.dataset.contactPoint && (
                                         <ContactPoint
                                             contactPoint={
@@ -519,7 +521,9 @@ class RecordHandler extends React.Component {
 
         return (
             <ReactDocumentTitle
-                title={`${title} | ${type} | ${config.appName}`}
+                title={`${title} | ${type} | ${
+                    this.props.strings.applicationName
+                }`}
             >
                 <div>{this.renderByState()}</div>
             </ReactDocumentTitle>
@@ -542,7 +546,8 @@ function mapStateToProps(state) {
         datasetIsFetching,
         distributionIsFetching,
         distributionFetchError,
-        datasetFetchError
+        datasetFetchError,
+        strings: state.content.strings
     };
 }